requirements.txt
cloudpickle==1.6.0
cycler==0.10.0
future==0.18.2
graphviz==0.16
gym==0.18.3
gym-retro==0.8.0
kiwisolver==1.3.1
matplotlib==3.4.2
neat-python==0.92
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.2.0
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.7.0
six==1.16.0
